Added MSVC CUDA project/solution
Added MSVC CUDA project/solution
Fixed warnings and errors when FLType is float
Note that when FLType is 32bit float, the maximum sigma with visually valid result is about 120
Format_Conversion.cpp removed
Format_Conversion.h renamed to Conversion.hpp, some template functions were moved from Image_Type.hpp
Some modification to class LUT (mainly cosmetics)

+++ include/Gaussian.h
@@ -81,6 +81,7 @@ public:
 
 
 // Implementation of recursive Gaussian algorithm from "Ian T. Young, Lucas J. van Vliet - Recursive implementation of the Gaussian filter"
+// For 32bit float type, the maximum sigma with valid result is about 120.
 class RecursiveGaussian
 {
 public:
@@ -93,7 +94,7 @@ protected:
     FLType B3;
 
 public:
-    _Myt(double sigma)
+    _Myt(long double sigma)
     {
         GetPara(sigma);
     }
@@ -105,7 +106,7 @@ public:
         return dst;
     }
 
-    void GetPara(double sigma);
+    void GetPara(long double sigma);
 
     void FilterV(Plane_FL &dst, const Plane_FL &src);
     void FilterV(Plane_FL &data);
