+++ source/Gaussian.cpp
@@ -23,15 +23,15 @@ Plane &Gaussian2D(Plane &dst, const Plane &src, const double sigma)
 
 
 // Member functions of class RecursiveGaussian
-void RecursiveGaussian::GetPara(double sigma)
+void RecursiveGaussian::GetPara(long double sigma)
 {
-    const double q = sigma < 2.5 ? 3.97156 - 4.14554 * sqrt(1 - 0.26891*sigma) : 0.98711 * sigma - 0.96330;
+    const long double q = sigma < 2.5 ? 3.97156 - 4.14554 * sqrt(1 - 0.26891 * sigma) : 0.98711 * sigma - 0.96330;
 
-    const double b0 = 1.57825 + 2.44413*q + 1.4281*q*q + 0.422205*q*q*q;
-    const double b1 = 2.44413*q + 2.85619*q*q + 1.26661*q*q*q;
-    const double b2 = -(1.4281*q*q + 1.26661*q*q*q);
-    const double b3 = 0.422205*q*q*q;
-
+    const long double b0 = 1.57825 + 2.44413*q + 1.4281*q*q + 0.422205*q*q*q;
+    const long double b1 = 2.44413*q + 2.85619*q*q + 1.26661*q*q*q;
+    const long double b2 = -(1.4281*q*q + 1.26661*q*q*q);
+    const long double b3 = 0.422205*q*q*q;
+    
     B = static_cast<FLType>(1 - (b1 + b2 + b3) / b0);
     B1 = static_cast<FLType>(b1 / b0);
     B2 = static_cast<FLType>(b2 / b0);
